refactor(bridge): replace GhCliService with API-layer GhCliApi, add GitRemoteApi
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
