+++ bridge/app/lib/src/bridge/api/gh_pull_request.freezed.dart
