+++ bridge/app/test/bridge/api/git_remote_api_test.dart
@@ -1,7 +1,7 @@
 import "dart:io";
 
-import "package:sesori_bridge/src/bridge/pr/remote_detection.dart";
-import "package:sesori_bridge/src/bridge/worktree_service.dart";
+import "package:sesori_bridge/src/bridge/api/git_remote_api.dart";
+import "package:sesori_bridge/src/bridge/foundation/process_runner.dart";
 import "package:test/test.dart";
 
 void main() {
@@ -12,8 +12,7 @@ void main() {
         stdout: "https://github.com/org/repo.git",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -26,8 +25,7 @@ void main() {
         stdout: "git@github.com:org/repo.git",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -40,8 +38,7 @@ void main() {
         stdout: "https://GitHub.COM/org/repo.git",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -54,8 +51,7 @@ void main() {
         stdout: "https://gitlab.com/org/repo.git",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -68,8 +64,7 @@ void main() {
         stdout: "/path/to/local/repo",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -82,8 +77,7 @@ void main() {
         stdout: "",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -96,8 +90,7 @@ void main() {
         stdout: "   \n  \t  ",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -110,8 +103,7 @@ void main() {
         stdout: "https://github.com/org/repo.git",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -128,8 +120,7 @@ void main() {
         return ProcessResult(0, 0, "https://github.com/org/repo.git", "");
       }
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -147,8 +138,7 @@ void main() {
         return ProcessResult(0, 0, "https://github.com/org/repo.git", "");
       }
 
-      await hasGitHubRemote(
-        processRunner: mockRunner,
+      await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/my/project/path",
       );
 
@@ -168,8 +158,7 @@ void main() {
         return ProcessResult(0, 0, "https://github.com/org/repo.git", "");
       }
 
-      await hasGitHubRemote(
-        processRunner: mockRunner,
+      await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -187,8 +176,7 @@ void main() {
         return ProcessResult(0, 127, "", "command not found");
       }
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
 
@@ -201,8 +189,7 @@ void main() {
         stdout: "  \n  https://github.com/org/repo.git  \n  ",
       );
 
-      final result = await hasGitHubRemote(
-        processRunner: mockRunner,
+      final result = await GitRemoteApi(processRunner: mockRunner).hasGitHubRemote(
         projectPath: "/path/to/project",
       );
